test: separate integration from unit test
